- Integrated new ListView system

+++ Linotte/CCGeohashHelper.m
@@ -44,12 +44,12 @@
     digits = MIN(digits, kCCGeohashHelperNDigits);
     NSUInteger power = kCCGeohashHelperNDigits - digits;
     NSUInteger digitsToMultiplier = pow(2, power);
-    for (int i = -radius; i <= radius; ++i) {
+    for (NSInteger i = -radius; i <= radius; ++i) {
         
-        for (int j = -radius; j <= radius; ++j) {
+        for (NSInteger j = -radius; j <= radius; ++j) {
             
-            if (all || !((!i && !j) || ((i == j || i == -j) && abs(i) == radius))) {
-                CCGeohashStruct geohash = init_neighbour(&centerGeohash, j * digitsToMultiplier, i * digitsToMultiplier);
+            if (all || !((!i && !j) || ((i == j || i == -j) && abs((int)i) == radius))) {
+                CCGeohashStruct geohash = init_neighbour(&centerGeohash, (int)(j * digitsToMultiplier), (int)(i * digitsToMultiplier));
                 NSString *hash = @(geohash.hash);
                 [geohashes addObject:hash];
                 //NSLog(@"i: %d  j: %d  geohash: %@  lat: %f  lng: %f", i, j, hash, geohash.latitude, geohash.longitude);
